[BE] Feat : Spring Security 구현

+++ server/routine/src/main/java/com/seb45_main_031/routine/member/service/MemberService.java
@@ -1,5 +1,7 @@
 package com.seb45_main_031.routine.member.service;
 
+import com.seb45_main_031.routine.auth.jwt.JwtTokenizer;
+import com.seb45_main_031.routine.auth.utils.CustomAuthorityUtils;
 import com.seb45_main_031.routine.exception.BusinessLogicException;
 import com.seb45_main_031.routine.exception.ExceptionCode;
 import com.seb45_main_031.routine.member.entity.Member;
@@ -7,19 +9,28 @@ import com.seb45_main_031.routine.member.repository.MemberRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @Transactional
 public class MemberService {
-
     private final MemberRepository memberRepository;
+    private final CustomAuthorityUtils customAuthorityUtils;
+    private final PasswordEncoder passwordEncoder;
+    private final JwtTokenizer jwtTokenizer;
 
-    public MemberService(MemberRepository memberRepository) {
+    public MemberService(MemberRepository memberRepository,
+                         CustomAuthorityUtils customAuthorityUtils,
+                         PasswordEncoder passwordEncoder,
+                         JwtTokenizer jwtTokenizer) {
         this.memberRepository = memberRepository;
+        this.customAuthorityUtils = customAuthorityUtils;
+        this.passwordEncoder = passwordEncoder;
+        this.jwtTokenizer = jwtTokenizer;
     }
 
     private Member findverifiedMember(long memberId){
@@ -39,14 +50,70 @@ public class MemberService {
         }
     }
 
+    private void checkMemberId(long memberId, String accessToken){
+
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+        long findMemberId = jwtTokenizer.getMemberIdFromAccessToken(accessToken, base64EncodedSecretKey);
+
+        if(memberId != findMemberId){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCHED);
+        }
+    }
+
+    private long verifyRefreshToken(String refreshToken, String base64EncodedSecretKey){
+
+        long findMemberId = jwtTokenizer.getMemberIdFromRefreshToken(refreshToken, base64EncodedSecretKey);
+
+        return findMemberId;
+    }
+
+    private String delegateAccessToken(long findMemberId, String base64EncodedSecretKey){
+
+        Member findMember = findverifiedMember(findMemberId);
+
+        Map<String, Object> claims = new HashMap<>();
+        claims.put("memberId", findMember.getMemberId());
+        claims.put("username", findMember.getEmail());
+        claims.put("roles", findMember.getRoles());
+
+        String subject = findMember.getEmail();
+
+        Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
+
+        String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
+
+        return accessToken;
+
+    }
+
 
     public Member createMember(Member member){
 
         verifyExistsEmail(member.getEmail());
 
+        String encryptedPassword = passwordEncoder.encode(member.getPassword());
+        member.setPassword(encryptedPassword);
+
+        List<String> roles = customAuthorityUtils.createRoles(member.getEmail());
+
+        member.setRoles(roles);
+
         return memberRepository.save(member);
 
     }
+
+    public String renewAccessToken(String refreshToken){
+
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+
+        long findMemberId = verifyRefreshToken(refreshToken, base64EncodedSecretKey);
+
+        String accessToken = delegateAccessToken(findMemberId, base64EncodedSecretKey);
+
+        return "Bearer " + accessToken;
+
+    }
+
 
     public Member updateMember(Member member){
         Member findMember = findverifiedMember(member.getMemberId());
@@ -58,22 +125,29 @@ public class MemberService {
 
     }
 
-    public Member findMember(long memberId){
-        return findverifiedMember(memberId);
+    public Member findMember(long memberId, String accessToken){
+
+        Member findMember = findverifiedMember(memberId);
+
+        checkMemberId(findMember.getMemberId(), accessToken);
+
+        return findMember;
     }
 
     public Page<Member> findMembers(int page, int size){
         return memberRepository.findAll(
                 PageRequest.of(page, size, Sort.by("memberId").descending()));
-
     }
 
-    public void deleteMember(long memberId){
+    public void deleteMember(long memberId, String password){
         Member findMember = findverifiedMember(memberId);
+
+        if(!passwordEncoder.matches(password, findMember.getPassword())) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PASSWORD_NOT_MATCHED);
+        }
 
         memberRepository.delete(findMember);
 
     }
-
 
 }
